requirements.txt
Flask
google-generativeai
gunicorn==20.1.0